--- Custom/events/SlaveAutoconf/README.rst
@@ -1,6 +1,0 @@
-SlaveAutoconf.py
-==================================
-
-This script written originally by Gavin Greenwalt and mangled by Edwin Amsler
-shows how to have new Slaves automatically configure a set of pools each time
-they start up.
