--- Custom/events/SlaveAutoconf/SlaveAutoconf.param
@@ -1,7 +1,0 @@
-[State]
-Type=Enum
-Items=Global Enabled;Opt-In;Disabled
-Label=State
-Default=Disabled
-Description=How this event plug-in should respond to events. If Global, all jobs and slaves will trigger the events for this plugin. If Opt-In, jobs and slaves can choose to trigger the events for this plugin. If Disabled, no events are triggered for this plugin.
-
