+++ Custom/events/WorkerAutoconf/README.rst
@@ -1,0 +1,6 @@
+WorkerAutoconf.py
+==================================
+
+This script written originally by Gavin Greenwalt and mangled by Edwin Amsler
+shows how to have new Slaves automatically configure a set of pools each time
+they start up.
